requirements.txt
Django>=5.1.6
djangorestframework
psycopg2-binary
djangorestframework-simplejwt